Add files via upload
There were a plethora of issues preventing the completion of this assignment. No colored texts that prevented any kind of tagging (this was solved later by identifying on the bottom-right page and switching it from Plain Text to HTML) as well as the inability of coding the images onto Sublime.Text and popping it up on the browser.